% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TWCR.annual.mean.hourly}
\alias{TWCR.annual.mean.hourly}
\title{Make annual mean fields for each hour.}
\usage{
TWCR.annual.mean.hourly(variable, year, version = 2)
}
\arguments{
\item{variable}{20CR variable name, only 2d variables will work.}

\item{year}{Year to make averages for.}

\item{version}{20CR version number - defaults to 2.}
}
\value{
A list (indexed 1:24) of fields, each of which is the annual average for that hour.
}
\description{
Poor man's normals - just the mean for each hour (1:24) over a year
}
\details{
Note that the indices are 1:24 not 0:23 as 0 can't be used as a list index.
}

