% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.TWCR.R
\name{TWCR.get.fixed.field}
\alias{TWCR.get.fixed.field}
\title{TWCR get fixed field.}
\usage{
TWCR.get.fixed.field(variable)
}
\arguments{
\item{variable}{- one of hgt.sfc, land, lsmask, soiltype, vegtype,}
}
\value{
field of selected data.
}
\description{
Get a field of non-time-varying data.
}
\details{
Currently doesn't vary betwen versions.
}
