% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TWCR.get.members.slab.from.hourly}
\alias{TWCR.get.members.slab.from.hourly}
\title{Extract a hyperslab of data - from individual ensemble members}
\usage{
TWCR.get.members.slab.from.hourly(variable, date.range, height.range = NULL,
  lat.range = c(-90, 90), lon.range = c(0, 360), members.range = c(0, 56),
  opendap = NULL, version = 2)
}
\arguments{
\item{variable}{20CR variable name, only 2d variables will work.}

\item{date.range}{A pair of text date strings, e.g. c('1981-02-05:12','1981-03-27:18') - inclusive.}

\item{height.range}{Bottom and top heights in hPa - leave NULL for monolevel}

\item{lat.range}{Min and max latitude in degrees - defaults to c(-90,90)}

\item{lon.range}{Min and max longitude in degrees - defaults to c(0,360)}

\item{members.range}{Min and max ensemble member - defaults to c(0,56)}

\item{opendap}{TRUE for network retrieval, FALSE for local files (faster, if you have them), NULL (default)
will use local files if available and network otherwise.}

\item{version}{20CR version number - defaults to 2.}

\item{type}{- 'mean' (default), 'spread', 'normal', or 'standard.deviation'.
Note that standard deviations are not available over opendap.}
}
\value{
A GSDF field with the selected multidimensional data
}
\description{
Up to 5d (lat, lon, height, time, member).
}

