% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TWCR.get.members.slice.at.hour}
\alias{TWCR.get.members.slice.at.hour}
\title{Get the ensemble members instead of the mean, spread etc.
analogous to TWCR.get.slice.at.hour}
\usage{
TWCR.get.members.slice.at.hour(variable, year, month, day, hour)
}
\arguments{
\item{variable}{must be 'prmsl'}
}
\value{
A GSDF field with ensemble number, lat and long as extended dimensions
}
\description{
Currently only for prmsl, and only for V2.
}
\details{
No interpolation - must be at an analysis time
}

