% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.TWCR.R
\name{TWCR.get.obs.1file}
\alias{TWCR.get.obs.1file}
\title{Get the observations from 1 prepbufr file}
\usage{
TWCR.get.obs.1file(year, month, day, hour, version = 2)
}
\value{
A data frame - one row for each observation.
}
\description{
All the observations used in one analysis run
 (0,6,12,18 hours each day).
}
\details{
Specification of obs format is at
http://rda.ucar.edu/datasets/ds131.1/docs/ISPD_quick_assimilated_ascii_format.pdf
file access only - the observations feedback files are not online.
}
