% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.TWCR.R
\name{TWCR.get.obs.1file.v3}
\alias{TWCR.get.obs.1file.v3}
\title{Get the observations from 1 psobfile file (v3 obs format)}
\usage{
TWCR.get.obs.1file.v3(year, month, day, hour, version = "4.1.8")
}
\value{
A data frame - one row for each observation.
}
\description{
All the observations used in one analysis run
 (0,6,12,18 hours each day).
}
