% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.TWCR.R
\name{TWCR.hourly.get.file.name}
\alias{TWCR.hourly.get.file.name}
\title{TWCR get file name (hourly)}
\usage{
TWCR.hourly.get.file.name(variable, year, month, day = NULL, hour = NULL,
  version = "3.5.1", type = "ensemble")
}
\arguments{
\item{variable}{'prmsl', 'prate', 'air.2m', 'uwnd.10m' or 'vwnd.10m' - or any 20CR variable}

\item{type}{- 'ensemble', 'mean', 'spread', 'normal', or 'standard.deviation'.
Also 'first.guess.mean' and 'first.guess.spread' for first guess fields.}
}
\value{
File name or URL for netCDF file containing the requested data
}
\description{
Get the file name (or URL) for selected variable and date (hourly data)
}
\details{
Called internally by \code{TWCR.get.slice.at.hour} but also useful
 called directly - you can then access the data with another tool.
}
