% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TWCR.hourly.members.get.file.name}
\alias{TWCR.hourly.members.get.file.name}
\title{TWCR get file name (ensemble members, hourly)}
\usage{
TWCR.hourly.members.get.file.name(variable, year, month, day, hour,
  opendap = NULL, version = 2)
}
\arguments{
\item{variable}{'prmsl', 'air.2m', prate, uwnd.10m and vwnd.10m. Others not currently available}

\item{opendap}{TRUE for network retrieval, FALSE for local files (faster, if you have them),
NULL (default) will use local files if available, and network if not.}
}
\value{
File name or URL for netCDF file containing the requested data
}
\description{
Get the file name (or URL) for selected variable and date (ensemble members hourly data)
}
\details{
Called internally by \code{TWCR.get.members.slice.at.hour} but also useful
 called directly - you can then access the data with another tool.
}

