% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatology.R
\name{TWCR.make.climatology}
\alias{TWCR.make.climatology}
\title{Make (and store) a climatological average}
\usage{
TWCR.make.climatology(variable, month = NULL, version, first.year = 1981,
  last.year = 2010)
}
\arguments{
\item{variable}{- 'air.2m', 'prmsl', 'prate', - 20CR names}

\item{month}{- 1-12, month to make climatology for.}

\item{version}{- 
If NULL (default) run for all months}
}
\value{
Nothing - a climatological field will be created as a side effect.
}
\description{
Data are stored as .nc files in a parallel directory to the original
 data - same as original data exept only for 1 year.
Loads a whole month of data at once
}
\details{
Chose to use 1981 as the storage year in the netcdf file.
}
