% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TWCR.members.bootstrap}
\alias{TWCR.members.bootstrap}
\title{Make a bootstrap version of an ensemble.}
\usage{
TWCR.members.bootstrap(ensemble)
}
\arguments{
\item{ensemble}{A GSDF field from TWCR.get.members.slice.at.hour}
}
\value{
A similar field, but with a bootstrap sample of the ensemble members
}
\description{
Extract the ensemble with TWCR.get.members.slice.at.hour
}
\details{
Makes a new ensemble subsampled with replacement from
 the original.
}

