% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TWCR.get.fixed.field}
\alias{TWCR.get.fixed.field}
\title{TWCR get fixed field.}
\usage{
TWCR.get.fixed.field(variable)
}
\arguments{
\item{variable}{- one of hgt.sfc, land, lsmask, soiltype, vegtype,}
}
\value{
field of selected data.
}
\description{
Get a field of non-time-varying data.
}
\details{
Always uses openDAP. Currently doesn't vary betwen versions.
}

