% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TWCR.get.members.slice.at.hour}
\alias{TWCR.get.members.slice.at.hour}
\title{Get the ensemble members instead of the mean, spread etc.
analogous to TWCR.get.slice.at.hour}
\usage{
TWCR.get.members.slice.at.hour(variable, year, month, day, hour,
  opendap = NULL, version = "2c")
}
\arguments{
\item{variable}{20CR variable name, only 'prmsl', 'air.2m', 'prate', 'uwnd.10m', and 'vwnd.10m' supported.}

\item{opendap}{TRUE for network retrieval, FALSE for local files (faster, if you have them).}
}
\value{
A GSDF field with ensemble number, lat and long as extended dimensions
}
\description{
Currently only some 2d variables.
}
\details{
No interpolation - must be at an analysis time
}

